Refactor geometry management
