+++ Managers/GeometryConversion.cs
@@ -4,26 +4,13 @@ using System.IO;
 using System.Linq;
 using Microsoft.Extensions.Logging;
 using Rhino;
-using Rhino.Commands;
 using Rhino.DocObjects;
 using Rhino.Geometry;
-using Rhino.Input;
-using Axys;
-using System;
-using System.Collections.Generic;
-using System.IO;
-
 
 namespace Axys
 {
-    public class GeometryManager
+    public static class GeometryConversion
     {
-
-        // Converts a PolyCurve into a joined Brep pipe. Each segment is piped and joined together.
-        // polyCurve - The PolyCurve to convert.
-        // doc - The active Rhino document, used for tolerance settings.
-        // pipeRadius - The radius to use for pipe creation.
-        // returns - A joined Brep representing the piped curve, or null if conversion failed.
         public static Brep ConvertPolyCurveToBrep(PolyCurve polyCurve, RhinoDoc doc, double pipeRadius)
         {
             var breps = new List<Brep>();
@@ -44,7 +31,6 @@ namespace Axys
                 {
                     breps.AddRange(pipeBreps);
                     Logger.LogDebug($"Created {pipeBreps.Length} pipe breps from subcurve.");
-
                 }
                 else
                 {
@@ -111,7 +97,6 @@ namespace Axys
 
         public static GeometryBase PrepareGeometryForExport(RhinoDoc doc, RhinoObject selectedObj)
         {
-
             GeometryBase geometry = selectedObj.Geometry?.Duplicate();
             if (geometry == null)
             {
@@ -274,268 +259,6 @@ namespace Axys
             }
 
             return geometry;
-        }
-
-        public static Guid AddTemporaryGeometryToDoc(RhinoDoc doc, GeometryBase geometry)
-        {
-            if (geometry == null)
-            {
-                Logger.LogError("Failed to prepare geometry for export.");
-                return Guid.Empty;
-            }
-
-            // 1. Compute the geometry's bounding box in its current world-space.
-            var bbox = geometry.GetBoundingBox(true);
-            // 2. Compute translation that moves the BOX CENTER to the origin
-            var baseCenter = new Point3d(
-                (bbox.Min.X + bbox.Max.X) / 2.0,
-                (bbox.Min.Y + bbox.Max.Y) / 2.0,
-                bbox.Min.Z
-            );
-            var offset = new Vector3d(-baseCenter.X, -baseCenter.Y, -baseCenter.Z);
-            var xformToOrigin = Transform.Translation(offset);
-
-            // 3. Apply translation
-            geometry.Transform(xformToOrigin);
-            Logger.LogDebug($"Geometry translated by {offset} to place its center at origin.");
-
-            Guid copyId = Guid.Empty;
-            if (geometry is Brep brep)
-            {
-                copyId = doc.Objects.AddBrep(brep);
-            }
-            else if (geometry is Mesh mesh)
-            {
-                copyId = doc.Objects.AddMesh(mesh);
-            }
-
-            if (copyId == Guid.Empty)
-            {
-                Logger.LogWarning("Temporary copy creation failed. Attempting export with original geometry.");
-            }
-            else
-            {
-                Logger.LogInfo("Temporary object added for export. Proceeding...");
-            }
-            return copyId;
-        }
-        //Needs testing, it's not clear if this is the right way to delete the temporary object
-        public static void DeleteTemporaryGeometry(RhinoDoc doc, Guid copyId)
-        {
-            if (copyId != Guid.Empty)
-            {
-                var obj = doc.Objects.Find(copyId);
-                if (obj != null)
-                {
-                    doc.Objects.Delete(obj, true);
-                    Logger.LogDebug($"Temporary object {copyId} deleted.");
-                }
-                else
-                {
-                    Logger.LogWarning($"Temporary object {copyId} not found for deletion.");
-                }
-            }
-            else
-            {
-                Logger.LogDebug("No temporary object to delete.");
-            }
-        }
-
-
-
-        public struct ExportResult
-        {
-            public bool Success;
-            public string Path;
-            public Guid TemporaryCopyId;
-        }
-        public static ExportResult ExportSelectedObjectToUSDZ(RhinoDoc doc, GeometryBase preparedGeometry, Guid objectId)
-        {
-            Guid exportId = Guid.NewGuid();
-
-            // Save the exported file in a dedicated temp directory
-            string tempDir = Path.Combine(Path.GetTempPath(), "RhinoExportTemp");
-            Directory.CreateDirectory(tempDir);
-            string fileName = $"Object_{objectId}.usdz";
-            string path = Path.Combine(tempDir, fileName);
-            Logger.LogDebug($"File will be saved as: {fileName}");
-            Logger.LogDebug($"Exporting object {objectId} to temporary path: {path}");
-            ExportToVision.LastExportedUSDZPath = path;
-
-            // Origin calculation block
-            Point3d origin;
-            var rhinoObj = doc.Objects.Find(objectId);
-            if (rhinoObj is InstanceObject)
-            {
-                origin = CalculateBlockInstanceOrigin
-    (rhinoObj);
-                Logger.LogDebug($"Origin from block instance center: {origin}");
-            }
-            else
-            {
-                var bbox = preparedGeometry.GetBoundingBox(true);
-                double centerX = (bbox.Min.X + bbox.Max.X) / 2.0;
-                double centerY = (bbox.Min.Y + bbox.Max.Y) / 2.0;
-                double minZ = double.MaxValue;
-                var vertices = new List<Point3d>();
-
-                if (preparedGeometry is Brep brep)
-                {
-                    foreach (var v in brep.DuplicateVertices())
-                        vertices.Add(v);
-                }
-                else if (preparedGeometry is Mesh mesh)
-                {
-                    vertices.AddRange(mesh.Vertices.ToPoint3dArray());
-                }
-
-                if (vertices.Count > 0)
-                {
-                    minZ = vertices.Min(v => v.Z);
-                }
-
-                origin = new Point3d(centerX, centerY, minZ);
-                Logger.LogDebug($"Origin from geometry bounding box: {origin}");
-            }
-
-            string originString = $"{origin.X},{origin.Y},{origin.Z}";
-
-            doc.Objects.UnselectAll();
-            doc.Objects.Select(objectId);
-            doc.Views.Redraw();
-
-            string script = string.Format(
-                "_-ExportWithOrigin {0} \"{1}\" _Enter _Enter",
-                originString, path);
-
-            Logger.LogDebug($"Export script command: {script}");
-            RhinoApp.RunScript(script, false);
-
-            bool fileExists = File.Exists(path);
-            if (fileExists)
-            {
-                Logger.LogDebug($"File found after export: {path}");
-            }
-            else
-            {
-                Logger.LogWarning($"File NOT found after export: {path}");
-            }
-
-            return new ExportResult
-            {
-                Success = fileExists,
-                Path = path,
-                TemporaryCopyId = Guid.Empty
-            };
-        }
-
-        public static Point3d CalculateBlockInstanceOrigin(RhinoObject obj)
-        {
-            if (!(obj is InstanceObject instanceObj))
-            {
-                Logger.LogWarning("Selected object is not an block.");
-                return Point3d.Unset;
-            }
-
-            var instanceDef = instanceObj.InstanceDefinition;
-            if (instanceDef == null)
-            {
-                Logger.LogWarning("Impossible to find the instance definition for the selected block.");
-                return Point3d.Unset;
-            }
-
-            var geometryTrasformate = new List<GeometryBase>();
-            foreach (var objId in instanceDef.GetObjectIds())
-            {
-                var rhinoObj = instanceObj.Document.Objects.Find(objId);
-                if (rhinoObj == null) continue;
-
-                var geo = rhinoObj.Geometry?.Duplicate();
-                if (geo == null) continue;
-
-                geo.Transform(instanceObj.InstanceXform);
-                geometryTrasformate.Add(geo);
-            }
-
-            if (geometryTrasformate.Count == 0)
-            {
-                Logger.LogWarning("No geometry found in the block instance definition.");
-                return Point3d.Unset;
-            }
-
-            BoundingBox bboxCombination = geometryTrasformate[0].GetBoundingBox(true);
-            for (int i = 1; i < geometryTrasformate.Count; i++)
-            {
-                bboxCombination.Union(geometryTrasformate[i].GetBoundingBox(true));
-            }
-
-            return bboxCombination.Center;
-        }
-
-        public class ScriptPipeMeshBlockCommand : Command
-        {
-            public static ScriptPipeMeshBlockCommand Instance { get; private set; }
-
-            public override string EnglishName => "ScriptPipeMeshBlock";
-
-            protected override Result RunCommand(RhinoDoc doc, RunMode mode)
-            {
-                // Select curves
-                ObjRef[] objRefs;
-                Result rc = RhinoGet.GetMultipleObjects(
-                    "Select curves to pipe, mesh, and block",
-                    false,
-                    ObjectType.Curve,
-                    out objRefs);
-
-                if (rc != Result.Success || objRefs == null || objRefs.Length == 0)
-                {
-                    Logger.LogWarning("No valid curves selected.");
-                    return Result.Cancel;
-                }
-
-                // Clear previous selections
-                doc.Objects.UnselectAll();
-
-                // Select the chosen curves
-                foreach (var o in objRefs)
-                    doc.Objects.Select(o.ObjectId);
-
-                doc.Views.Redraw();
-
-                // Pipe each curve
-                RhinoApp.RunScript("-_Pipe 0.3 0.3 _Enter _Enter", false);
-
-                // Mesh
-                RhinoApp.RunScript("-_Mesh _Enter _Enter", false);
-
-                // Retrieve the last created mesh objects
-                int expectedMeshes = objRefs.Length;
-                var meshObjs = doc.Objects.GetObjectList(ObjectType.Mesh)
-                    .OrderByDescending(o => o.RuntimeSerialNumber)
-                    .Take(expectedMeshes)
-                    .ToList();
-
-                if (meshObjs.Count == 0)
-                {
-                    Logger.LogError("No meshes found for block creation.");
-                    return Result.Failure;
-                }
-
-                // Select the created meshes
-                doc.Objects.UnselectAll();
-                foreach (var m in meshObjs)
-                    doc.Objects.Select(m.Id);
-
-                // Create a block from the selected meshes
-                var basePt = new Point3d(0, 0, 0);
-                string idList = string.Join(" ", meshObjs.Select(m => m.Id));
-                RhinoApp.RunScript($"-_Block {idList} {basePt.X},{basePt.Y},{basePt.Z} _Enter", false);
-
-                Logger.LogDebug("Pipe → Mesh → Block pipeline completed.");
-
-                return Result.Success;
-            }
         }
     }
 }
